+++ ExcelToNLS/CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2022-11-29
+
+### Added
+
+- Add a XLSX to XML converter.
+- Add a XLSX to XML launcher.
+
+### Changed
+
+- Project structure (separate XML and PO specific classes).
+
+### Fixed
+
+- Fix typo in the translations.xlsx file.
+
 ## [3.0.0] - 2021-09-13
 
 ### Added
@@ -33,5 +48,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 _Markdown_   
-_Copyright 2021 MicroEJ Corp. All rights reserved._  
+_Copyright 2021-2022 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._