+++ HelloWorldNLS/CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.0] - 2022-11-29
+
+### Changed 
+
+- Use ``binary-nls#2.5.0`` instead of ``nls-po#2.2.0`` api.
+
+### Added
+
+- Support of [string resources](https://developer.android.com/guide/topics/resources/string-resource) for translation file.
+- Update example to use both po and xml files.
+- Add fallback example.
+
 ## [3.0.0] - 2021-09-13
 
 ### Added
@@ -45,5 +57,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 _Markdown_   
-_Copyright 2021 MicroEJ Corp. All rights reserved._  
+_Copyright 2021-2022 MicroEJ Corp. All rights reserved._  
 _Use of this source code is governed by a BSD-style license that can be found with this software._